+++ ProjecTree-web/src/apis/index.ts
@@ -1,3 +1,4 @@
+// ===== Auth =====
 export {
   redirectToGithubOauth,
   redirectToGoogleOauth,
@@ -6,8 +7,10 @@ export {
   refreshToken,
 } from './oauth.api';
 
+// ===== Chat =====
 export { fetchMessages, fetchParticipants } from './chat.api';
 
+// ===== Member =====
 export {
   checkNicknameDuplicate,
   deleteMember,
@@ -17,6 +20,9 @@ export {
   updateNickname,
 } from './member.api';
 
+export type { MemberInfoResponse } from './member.api';
+
+// ===== Node =====
 export {
   getAiNodeTechRecommendation,
   postCreateCustomNode,
@@ -24,10 +30,38 @@ export {
   postCustomNodeTechRecommendation,
 } from './node.api';
 
+// ===== Portfolio =====
 export {
   generatePortfolio,
   getPortfolio,
   updatePortfolio,
 } from './portfolio.api';
 
-export { changeMemberRole, inviteMember } from './team.api';
+// ===== Team =====
+export { changeMemberRole, inviteMember, TeamRole, getRoleLabel } from './team.api';
+
+export type { TeamRoleType } from './team.api';
+
+// ===== Workspace =====
+export {
+  getWorkspaceTree,
+  getNodeDetail,
+  generateNodeCandidates,
+  selectNodeCandidates,
+  getWorkspaceDetail,
+  updateWorkspace,
+  createWorkspace,
+  getTechStacks,
+} from './workspace.api';
+
+export type {
+  WorkspaceDetailData,
+  Role,
+  TechStackItem,
+  CreateWorkspaceFormData,
+} from './workspace.api';
+
+// ===== Workspace Lounge =====
+export { getMyWorkspaces } from './workspace-lounge.api';
+
+export type { WorkspaceCardData } from './workspace-lounge.api';
